// src/models/Enums.ts
export const enum Category {
  Tv = "tv",
  Movie = "movie",
}

export const enum MovieType {
  Popular = "popular",
  TopRated = "top_rated",
  Similar = "similar",
  Upcoming = "upcoming",
}

export const enum TvType {
  Popular = "popular",
  TopRated = "top_rated",
  Similar = "similar",
  OnTheAir = "on_the_air",
}